# clients/go-client/api/openapi.yaml
openapi: 3.1.0
info:
  title: Platform API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /ada/v1/projects:
    get:
      operationId: list_projects_ada_v1_projects_get
      parameters:
        - explode: true
          in: query
          name: skip
          required: false
          schema:
            default: 0
            title: Skip
            type: integer
          style: form
        - explode: true
          in: query
          name: limit
          required: false
          schema:
            default: 100
            title: Limit
            type: integer
          style: form
        - explode: false
          in: header
          name: x-ada-token
          required: false
          schema:
            title: X-Ada-Token
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
        - HTTPBearer: []
      summary: List Projects
      tags:
        - projects
    post:
      operationId: create_ada_v1_projects_post
      parameters:
        - explode: false
          in: header
          name: x-ada-token
          required: false
          schema:
            title: X-Ada-Token
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
        - HTTPBearer: []
      summary: Create
      tags:
        - projects
  /ada/v1/projects/{project_id}/chats:
    get:
      operationId: read_chats_ada_v1_projects__project_id__chats_get
      parameters:
        - explode: false
          in: path
          name: project_id
          required: true
          schema:
            title: Project Id
            type: integer
          style: simple
        - explode: true
          in: query
          name: skip
          required: false
          schema:
            default: 0
            title: Skip
            type: integer
          style: form
        - explode: true
          in: query
          name: limit
          required: false
          schema:
            default: 100
            title: Limit
            type: integer
          style: form
        - explode: false
          in: header
          name: x-ada-token
          required: false
          schema:
            title: X-Ada-Token
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Chat"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
        - HTTPBearer: []
      summary: Read Chats
      tags:
        - projects
    post:
      operationId: create_chat_ada_v1_projects__project_id__chats_post
      parameters:
        - explode: false
          in: path
          name: project_id
          required: true
          schema:
            title: Project Id
            type: integer
          style: simple
        - explode: false
          in: header
          name: x-ada-token
          required: false
          schema:
            title: X-Ada-Token
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCreate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
        - HTTPBearer: []
      summary: Create Chat
      tags:
        - projects
...
components:
  schemas:
...
    Chat:
      example:
        project_id: 0
        name: name
        model: model
        id: 6
      properties:
        name:
          title: Name
          type: string
        model:
          title: Model
          type: string
        project_id:
          title: Project Id
          type: integer
        id:
          title: Id
          type: integer
      required:
        - id
        - model
        - name
        - project_id
      title: Chat
    ChatCreate:
      example:
        project_id: 0
        name: name
        model: model
      properties:
        name:
          title: Name
          type: string
        model:
          title: Model
          type: string
        project_id:
          title: Project Id
          type: integer
      required:
        - model
        - name
        - project_id
      title: ChatCreate
...
    HTTPValidationError:
      example:
        detail:
          - msg: msg
            loc:
              - ValidationError_loc_inner
              - ValidationError_loc_inner
            type: type
          - msg: msg
            loc:
              - ValidationError_loc_inner
              - ValidationError_loc_inner
            type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    ProjectCreate:
      example:
        organization_id: 0
        name: name
      properties:
        name:
          title: Name
          type: string
        organization_id:
          nullable: true
          type: integer
      required:
        - name
      title: ProjectCreate
...
    ValidationError:
      example:
        msg: msg
        loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
        - loc
        - msg
        - type
      title: ValidationError
    ValidationError_loc_inner:
      anyOf:
        - type: string
        - type: integer
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http